Hit boxes?
Now anything that can be clicked on has a hit box member. The hit box has a unique call back function. The button still says hello, but now the states have the opprotunitiy of displaying their info in the console.

NOTE: some of the hitboxes seem to be broken as there are 801 on screen when there should only be 401. In the middle when you click, multiple hit boxes' callback functions are triggered. This may be an issue dealing with the amount of actual states were spawned in...